report_api/requirements.txt
# Requirements for Report API
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.7
ItsDangerous==1.0.0
Jinja2==2.10
MarkupSafe==1.0
python-dateutil==2.7.3
requests==2.20.0
six==1.11.0
urllib3==1.24
Werkzeug==0.14.1
